Story Pictures and Content
Add the story part of the project with images overlaping each other and some text on the side.

Use a grid to make a six column and row template. (Remember how to use it to make good layouts with it, think how many are necesary is good to make complex and beautiful styles).

I can use grid to center elements on the grid container or I can mix grid with flexbox to make the same operation.

Add few typography styles but nothing new.