Working with JavaScript
1. Created three HTML / CSS new features for the webpage;

1.1. Created and styled the NAVIGATION MENU with already used CSS-FLEX properties. Not many difficulties to style it but sometimes getting crazy with the heigth property when trying to make elements responsive.

Added hover effects that make a background transition from left to right. Used background-clip and text fill color properties. CHECK THIS AGAIN!!!

1.2. Created the MODAL FORM that is fixed when opened. Actually not many difficulties to style it. Thanks to the overflow property that is making things quite easy.

Learned the backdrop-property that blurs the background, but which is not working in firefox. FOr that used the @-moz-document url-prefix()
to make the style in Firefox different.

1.3. Created a GALLERY SLIDER MODAL for the images. Quite 'hard' to make the Interface properly styled. Getting really crazy with the heigth of the elements when working with responsive elements. And at last needed from JS to make the interface correct work as a normal slider.

2. Created four JAVASCRIPT new features for the webpage;

2.1. Created a NAVIGATION MENU that is displayed through a hamburger button styled to change appearance when clicked. Make the NavMenu show and hide when hamburger is clicked with JS.

To make it work 'Show and Hide' two function have been created. Both functions work in similar way. When hamburger is clicked the function make the hamburger change and at the same time toggles a configured CSS class to change the display property.

2.2 Created a function to show and hide a CONTACT FORM when specific  buttons are clicked. As well as in the last function when contact button is clicked the functions removes a CSS class.

At the same time the whole body is fixed to stop the scrolling at the background when form is opened, using the overflow = hidden property.

To make the form show when contact button is clicked is necessary to create a loop, which I DO NOT REALLY UNDERSTAND HOW IT WORKS or why is necessary in this case. CHECK!!!

The close function works similar, but in this case adds a CSS class, and returns the body to the default scroll property.

2.3 Created a function to toggle the HEART COLOR when clicked.

This function has similar features due to it works with the toggle attribute to  able and disable the like button when clicked.

2.4 Created a MODAL GALLERY that works as a slider with images. (NOT FINISHED)

The gallery contains the main picture that is centered in the screen with an arrow in each side to make navigate between the images.

First of all the images 'slides', are translate in X direction to make a large row with them. CHECK THIS PROPERTY AND UNDERSTAND HOW IT REALLY WORKS. '' forEach''.

To make the arrows works, created to events that make go to the next image.

Used the already aforementioned forEach attribute that makes the translateX property change everytime that arrows are clicked. CHECK!!!!!!!

To fix the bug of the translateX go to infinite, to make it work needed to create two variables to specify, 1. the current slide and 2. the max length of the images.

Inside the whole function create an if else statement that makes go to first slide if we are on the last one and the same other way around.

3. ------------------  RESUME AND CONSIDERATIONS ----------------------

3.1 I do not really have difficulties when styling new elements with CSS, I am confortable with the positioning and flex, and not using Grid so much. I think for specific things is much better.

3.2 About JavaScript, being a project to learn and reforce the fundamentals I being quite overwhelming behind the scenes. I mean there are times where I get confused with the syntax and how I should organize a JS file.

In general all the features used in this file I know how they work but I am sure that I am missing many things that will have a better solution.

Not sure about semicolons, where and when to declare variables when working with functions. NEED TO READ MORE ABOUT HOW JAVASCRIPT WORKS.

To be one of the first webpages where I use JS I am quite happy about the results, I am sure that many things are not perfect but I made things work as I want.

¡¡¡¡¡¡¡¡¡ Even if there are things that are not finished yet. !!!!!!!!

GENESIS MANDELA 11 NOV